+++ src/sparseml/modifiers/pruning/wanda/utils/layer_compressor.py
@@ -12,24 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import inspect
-import logging
-from typing import Dict, List
-
-import torch
-from torch.nn import Module
+from typing import Dict
 
 from sparseml.modifiers.pruning.wanda.utils.wandagpt import WandaGPT
-from sparseml.pytorch.utils.helpers import get_dependency_order
-from sparseml.utils.pytorch.module import get_prunable_layers
+from sparseml.modifiers.utils.gpt_helpers import TerminalModuleCompressor
+from sparseml.modifiers.utils.layer_compressor import LayerCompressor
 
 
-__all__ = ["WandaLayerCompressor"]
-
-_LOGGER = logging.getLogger(__name__)
-
-
-class WandaLayerCompressor:
+class WandaLayerCompressor(LayerCompressor):
     """
     Runs the Wanda algorithm on a single layer using calibration data inputs
 
@@ -47,63 +37,7 @@ class WandaLayerCompressor:
     :param args: additional keyword arguments
     """
 
-    def __init__(
-        self, model: Module, layer: Module, layer_index: int, inputs: List, args: Dict
-    ):
-        self.model = model
-        self.layer = layer
-        self.layer_index = layer_index
-        self.inputs = inputs
-        self.args = args
-
-    def compressible_modules(self) -> Dict:
-        """
-        Get the list of modules in the layer that can be compressed
-
-        :return: dictionary of compressible modules
-        """
-        compressible_layers = get_prunable_layers(self.layer)
-        return compressible_layers
-
-    def pre_compress_parallel(self, **kwargs) -> Dict:
-        """
-        Sets up the WrappedGPT objects for each compressible module,
-        computes the statistics for each using the calibration data.
-
-        :return: WrappedGPT objects for each module
-        """
-        subset = self.compressible_modules()
-
-        gpts = {}
-        for name in subset:
-            gpts[name] = WandaGPT(subset[name])
-
-        def add_batch(name):
-            def tmp(_, inp, out):
-                gpts[name].add_batch(inp[0].data, out.data)
-
-            return tmp
-
-        handles = []
-        for name in gpts:
-            handles.append(subset[name].register_forward_hook(add_batch(name)))
-
-        # Run through the samples in order to compute statistics for each module
-        nsamples = len(self.inputs)
-        forward_args_spec = inspect.getfullargspec(self.layer.__class__.forward)
-        passed_in_args = [arg for arg in forward_args_spec.args if arg in kwargs]
-        for sample_idx in range(nsamples):
-            passed_in_kwargs = {}
-            for arg in passed_in_args:
-                if isinstance(kwargs[arg], List):
-                    passed_in_kwargs[arg] = kwargs[arg][sample_idx]
-                else:
-                    passed_in_kwargs[arg] = kwargs[arg]
-            self.layer(self.inputs[sample_idx], **passed_in_kwargs)
-        for h in handles:
-            h.remove()
-
-        return {"gpts": gpts}
+    gpt_class: TerminalModuleCompressor = WandaGPT
 
     def compress(self, dev: str = "cuda:0", **kwargs) -> Dict:
         """
@@ -116,82 +50,10 @@ class WandaLayerCompressor:
         extras = self.post_compress(**kwargs)
         return {"outputs": extras["outputs"]}
 
-    def post_compress(self, **kwargs) -> Dict:
-        """
-        Clean up after compression
-
-        :return: outputs of the layer
-        """
-        nsamples = len(self.inputs)
-        outputs = []
-        forward_args_spec = inspect.getfullargspec(self.layer.__class__.forward)
-        passed_in_args = [arg for arg in forward_args_spec.args if arg in kwargs]
-        for j in range(nsamples):
-            passed_in_kwargs = {}
-            for arg in passed_in_args:
-                if isinstance(kwargs[arg], List):
-                    passed_in_kwargs[arg] = kwargs[arg][j]
-                else:
-                    passed_in_kwargs[arg] = kwargs[arg]
-            outputs.append(self.layer(self.inputs[j], **passed_in_kwargs)[0])
-
-        self.inputs = None
-        # once we've finished compressing the layer, move it back to CPU to save memory
-        self.layer.to("cpu")
-        torch.cuda.empty_cache()
-
-        return {"outputs": outputs}
-
-    def sequentially_compress(self, **kwargs):
-        """
-        Run compression module by module, in dependency order. Unlike in parallel
-        compression, we compute the statistics layer by layer instead of computing them
-        all up front. This saves on memory and means compression in earlier layers
-        affects the inputs to later layers
-        """
-        subset = self.compressible_modules()
-
-        # filter kwargs that are expected as layer inputs
-        forward_args_spec = inspect.getfullargspec(self.layer.__class__.forward)
-        passed_in_args = [arg for arg in forward_args_spec.args if arg in kwargs]
-
-        passed_in_kwargs = {}
-        for arg in passed_in_args:
-            if isinstance(kwargs[arg], List):  # take the first batch
-                passed_in_kwargs[arg] = kwargs[arg][0]
-            else:
-                passed_in_kwargs[arg] = kwargs[arg]
-        order = get_dependency_order(
-            self.layer, subset, self.inputs[0], **passed_in_kwargs
+    def invoke_fasterprune(self, gpts: WandaGPT):
+        # run WandaGPT algorithm on current module
+        gpts.fasterprune(
+            self.args["sparsity"],
+            prunen=self.args["prunen"],
+            prunem=self.args["prunem"],
         )
-
-        nsamples = len(self.inputs)
-        for name in order:  # create WrappedGPT object for each compressible module
-            gpts = WandaGPT(subset[name])
-
-            def add_batch(name):
-                def tmp(_, inp, out):
-                    gpts.add_batch(inp[0].data, out.data)
-
-                return tmp
-
-            # add WrappedGPT hook for current module
-            handle = subset[name].register_forward_hook(add_batch(name))
-            for sample_idx in range(nsamples):
-                passed_in_kwargs = {}
-                for arg in passed_in_args:
-                    if isinstance(kwargs[arg], List):
-                        passed_in_kwargs[arg] = kwargs[arg][sample_idx]
-                    else:
-                        passed_in_kwargs[arg] = kwargs[arg]
-                # run layer, triggering WrappedGPT add_batch for current module
-                self.layer(self.inputs[sample_idx], **passed_in_kwargs)
-            handle.remove()
-
-            _LOGGER.info(f"Compressing module {name} of layer {self.layer_index}")
-            gpts.fasterprune(  # run WrappedGPT algorithm on current module
-                self.args["sparsity"],
-                prunen=self.args["prunen"],
-                prunem=self.args["prunem"],
-            )
-            gpts.free()
